+++ CHANGES.rst
@@ -3,7 +3,8 @@ CHANGE HISTORY
 
 1.1.0
 -----
-- Added the Julia Set effect.
+- Added the Julia Set and Cog effects.
+- Fixed up off-by-one error in line drawing.
 
 1.0.0
 -----
@@ -16,7 +17,7 @@ CHANGE HISTORY
 0.4.0
 -----
 - Added support for 256 colour terminals.
-- Moved ${c,a} syntax for inline colouring from Screen to Renderers
+- Moved ${c,a} syntax for inline colouring from Screen to Renderer.
 - Created some samples for 256 colour mode and colour images.
 
 0.3.0